requirements.txt
openai
langchain
langchain_openai
langchain-community
chromadb
pinecone
python-dotenv
